justplucked/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
blis==1.3.0
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ChatterBot==1.2.6
chatterbot-corpus==1.2.2
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
crispy-bootstrap5==2025.4
cryptography==44.0.2
cymem==2.0.11
Django==5.2
django-allauth==65.7.0
django-countries==7.6.1
django-crispy-forms==2.4
django-jazzmin==3.0.1
django-tinymce==4.1.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
greenlet==3.2.1
idna==3.10
Jinja2==3.1.6
jwt==1.3.1
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mathparse==0.1.5
mdurl==0.1.2
murmurhash==1.0.12
numpy==2.2.5
packaging==25.0
pillow==11.2.1
preshed==3.0.9
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
requests==2.32.3
rich==14.0.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
sqlparse==0.5.3
srsly==2.5.1
stripe==12.0.1
thinc==8.3.6
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
